+++ testing/jormungandr-integration-tests/src/networking/p2p/stats.rs
@@ -1,16 +1,27 @@
 use crate::networking::utils;
+use chain_impl_mockchain::header::BlockDate;
+use jormungandr_lib::interfaces::NodeStats;
 use jormungandr_lib::interfaces::Policy;
+use jormungandr_lib::interfaces::SlotDuration;
+use jormungandr_testing_utils::testing::jormungandr::JormungandrProcess;
 use jormungandr_testing_utils::testing::network::builder::NetworkBuilder;
 use jormungandr_testing_utils::testing::network::wallet::template::builder::WalletTemplateBuilder;
+use jormungandr_testing_utils::testing::network::Blockchain;
 use jormungandr_testing_utils::testing::network::Node;
 use jormungandr_testing_utils::testing::network::SpawnParams;
 use jormungandr_testing_utils::testing::network::Topology;
+use jormungandr_testing_utils::testing::FragmentSender;
+use std::fmt::Display;
 use std::time::Duration;
 
 const LEADER1: &str = "LEADER1";
 const LEADER2: &str = "LEADER2";
 const LEADER3: &str = "LEADER3";
 const LEADER4: &str = "LEADER4";
+
+const PASSIVE: &str = "PASSIVE";
+const LEADER_CLIENT: &str = "LEADER_CLIENT";
+const LEADER: &str = "LEADER";
 
 const ALICE: &str = "ALICE";
 const BOB: &str = "BOB";
@@ -136,4 +147,190 @@ pub fn p2p_stats_test() {
     super::assert_node_stats(&leader1, 3, 0, 3, "leader1: leader 2 is down");
     super::assert_node_stats(&leader3, 3, 0, 3, "leader3: leader 2 is down");
     super::assert_node_stats(&leader4, 3, 0, 3, "leader4: leader 2 is down")
+}
+
+// build a blockchain with a longer slot duration than default
+// to avoid spurious failures as described in
+// https://github.com/input-output-hk/jormungandr/issues/3183.
+// It is a macro because the builder is returned by reference.
+macro_rules! build_network {
+    () => {{
+        let mut blockchain = Blockchain::default();
+        blockchain.set_slot_duration(SlotDuration::new(5).unwrap());
+        NetworkBuilder::default().blockchain_config(blockchain)
+    }};
+}
+
+#[test]
+pub fn passive_node_last_block_info() {
+    let mut network_controller = build_network!()
+        .topology(
+            Topology::default()
+                .with_node(Node::new(LEADER))
+                .with_node(Node::new(PASSIVE).with_trusted_peer(LEADER)),
+        )
+        .wallet_template(
+            WalletTemplateBuilder::new("alice")
+                .with(1_000_000)
+                .delegated_to(LEADER)
+                .build(),
+        )
+        .wallet_template(WalletTemplateBuilder::new("bob").with(1_000_000).build())
+        .build()
+        .unwrap();
+
+    let leader = network_controller
+        .spawn(SpawnParams::new(LEADER).in_memory())
+        .unwrap();
+    let passive = network_controller
+        .spawn(SpawnParams::new(PASSIVE).in_memory().passive())
+        .unwrap();
+
+    let mut alice = network_controller.wallet("alice").unwrap();
+    let mut bob = network_controller.wallet("bob").unwrap();
+
+    let stats_before = passive
+        .rest()
+        .stats()
+        .expect("cannot get stats at beginning")
+        .stats
+        .expect("empty stats");
+
+    let fragment_sender = FragmentSender::new(
+        leader.genesis_block_hash(),
+        leader.fees(),
+        BlockDate::first().next_epoch().into(),
+        Default::default(),
+    );
+
+    fragment_sender
+        .send_transactions_round_trip(5, &mut alice, &mut bob, &leader, 100.into())
+        .expect("fragment send error");
+
+    assert_last_stats_are_updated(stats_before, &passive);
+}
+
+#[test]
+pub fn leader_node_last_block_info() {
+    let mut network_controller = build_network!()
+        .topology(
+            Topology::default()
+                .with_node(Node::new(LEADER))
+                .with_node(Node::new(LEADER_CLIENT).with_trusted_peer(LEADER)),
+        )
+        .wallet_template(
+            WalletTemplateBuilder::new("alice")
+                .with(1_000_000)
+                .delegated_to(LEADER)
+                .build(),
+        )
+        .wallet_template(WalletTemplateBuilder::new("bob").with(1_000_000).build())
+        .build()
+        .unwrap();
+
+    let leader = network_controller
+        .spawn(SpawnParams::new(LEADER).in_memory())
+        .unwrap();
+    let leader_client = network_controller
+        .spawn(SpawnParams::new(LEADER_CLIENT).in_memory())
+        .unwrap();
+
+    let mut alice = network_controller.wallet("alice").unwrap();
+    let mut bob = network_controller.wallet("bob").unwrap();
+
+    let stats_before = leader_client
+        .rest()
+        .stats()
+        .expect("cannot get stats at beginning")
+        .stats
+        .expect("empty stats");
+
+    let fragment_sender = FragmentSender::new(
+        leader.genesis_block_hash(),
+        leader.fees(),
+        BlockDate::first().next_epoch().into(),
+        Default::default(),
+    );
+
+    fragment_sender
+        .send_transactions_round_trip(5, &mut alice, &mut bob, &leader, 100.into())
+        .expect("fragment send error");
+
+    assert_last_stats_are_updated(stats_before, &leader_client);
+}
+
+fn assert_last_stats_are_updated(stats_before: NodeStats, node: &JormungandrProcess) {
+    let stats_after = node
+        .rest()
+        .stats()
+        .expect("cannot get stats at end")
+        .stats
+        .expect("empty stats");
+
+    stats_element_is_different(
+        stats_before.last_block_content_size,
+        stats_after.last_block_content_size,
+        "last block content size",
+    );
+
+    let before_last_block_date: BlockDate = stats_before.last_block_date.unwrap().parse().unwrap();
+    let after_last_block_date: BlockDate = stats_after.last_block_date.unwrap().parse().unwrap();
+
+    stats_element_is_greater(
+        before_last_block_date,
+        after_last_block_date,
+        "last block date",
+    );
+    stats_element_is_different(
+        stats_before.last_block_fees,
+        stats_after.last_block_fees,
+        "last block fees size",
+    );
+
+    stats_element_is_different(
+        stats_before.last_block_hash.unwrap(),
+        stats_after.last_block_hash.unwrap(),
+        "last block hash",
+    );
+    stats_element_is_different(
+        stats_before.last_block_sum,
+        stats_after.last_block_sum,
+        "last block sum",
+    );
+    stats_element_is_greater(
+        stats_before.last_block_time.unwrap(),
+        stats_after.last_block_time.unwrap(),
+        "last block time",
+    );
+    stats_element_is_different(
+        stats_before.last_block_tx,
+        stats_after.last_block_tx,
+        "last block tx",
+    );
+}
+
+fn stats_element_is_greater<T>(before_value: T, after_value: T, info: &str)
+where
+    T: Display + PartialOrd,
+{
+    assert!(
+        before_value < after_value,
+        "{} should to be updated. {} vs {}",
+        info,
+        before_value,
+        after_value,
+    );
+}
+
+fn stats_element_is_different<T>(before_value: T, after_value: T, info: &str)
+where
+    T: Display + PartialOrd,
+{
+    assert!(
+        before_value != after_value,
+        "{} should to be updated. {} vs {}",
+        info,
+        before_value,
+        after_value,
+    );
 }
